ml/requirements.txt
python-dotenv
groq
langchain
langchain-community
langchainhub
langchain-groq
langchain_openai
pandas
numpy
moviepy
youtube_search
youtube-transcript-api
pytube
pytubefix
requests
audio-extract
faster-whisper
openai-whisper
huggingface_hub[hf_xet]
hf_xet
langchain-chroma
pandas
numpy 
matplotlib
scikit-learn
flask 
flask-cors
langchain-experimental
pillow 
open_clip_torch 
torch 
matplotlib
pypdf
pydub